pyproject.toml
[project]
name = "ml-blueprint"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "lightgbm>=4.6.0",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.7.1",
    "xgboost>=3.0.4",
]

[project.scripts]
ml-blueprint = "ml_blueprint:main"

[build-system]
requires = ["uv_build>=0.8.11,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]
